Refactors manage.py (#992)
* App and configure commands

* Athena command

* Terraform build and clean

* Create alarm commands

* Metrics command

* Fix bug

* Terraform commands

* List targets command

* Status command

* Kinesis and rollback command

* Rule stagnign

* Finished

* Fixups

* Ye

* Pylint

* Done

* PR feedback; cleanup

* Pr feedback
